GuardianBot/requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.12
distlib==0.3.4
docutils==0.18.1
filelock==3.6.0
idna==3.3
Kivy==2.1.0
Kivy-examples==2.1.0
Kivy-Garden==0.1.5
Pillow==9.1.0
platformdirs==2.5.2
Pygments==2.12.0
python-telegram-bot==13.11
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
six==1.16.0
tornado==6.1
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
virtualenv==20.14.1
